refactor(protocol-contracts): refactor logExplorerLink utils
