+++ protocol-contracts/token/tasks/roles/helpers.ts
@@ -1,8 +1,6 @@
 import { Contract } from 'ethers'
 import { HardhatRuntimeEnvironment } from 'hardhat/types'
 
-import { getBlockExplorerLink } from '../utils'
-
 export type RoleKey = 'MINTER_ROLE' | 'PAUSING_MINTER_ROLE' | 'DEFAULT_ADMIN_ROLE'
 
 export const ROLE_ACCESSORS: Record<RoleKey, string> = {
@@ -56,23 +54,4 @@ export async function resolveRoleValue(zamaErc20: Contract, role: RoleKey): Prom
         throw new Error(`Role accessor ${accessor}() not available on ZamaERC20 contract`)
     }
     return getter()
-}
-
-export async function logExplorerLink(hre: HardhatRuntimeEnvironment, txHash: string): Promise<void> {
-    const { eid } = hre.network.config as { eid?: number }
-    if (typeof eid !== 'number') {
-        console.log('No endpoint ID configured for this network; unable to derive block explorer link.')
-        return
-    }
-
-    try {
-        const explorerLink = await getBlockExplorerLink(eid, txHash)
-        if (explorerLink) {
-            console.log(`Block explorer: ${explorerLink}`)
-        } else {
-            console.log('Block explorer URL unavailable for this network; check LayerZero metadata service.')
-        }
-    } catch (error) {
-        console.log(`Failed to retrieve block explorer URL: ${error instanceof Error ? error.message : error}`)
-    }
 }
